requirements/requirements.txt
Django==3.2.18
psycopg2==2.9.5
djangorestframework==3.14.0
djoser==2.1.0
django-silk==5.0.3
django-cors-headers==3.14.0
python-slugify==8.0.1
Pillow==9.4.0